Do not get a Cipher instance every time we decrypt a chunk (#215)
Since we're calling cipher.init() in every call, there's no need to get the instance every time. Just save it as a private final variable.
In case it ever throws an exception during decoding, running ciper.init() again will bring it to a correct state again (according to the documentation).